Fix: Address PR #12 high-priority feedback
Fixes two high-priority issues identified by Codex review:

Issue #1: BigDecimal Precision Loss
- Removed .toDouble conversions in FormattedParsers (3 locations)
- parseMoney, parsePercent, parseAccounting now preserve full BigDecimal precision
- Added test to verify high-precision values (1234567890.123456789)

Issue #2: Formula Parser String Literals
- Enhanced validateParentheses to respect string literals ("text")
- Tracks quote state, ignores parens inside strings
- Handles escaped quotes ("") correctly per Excel convention
- Detects unclosed strings (returns false if inString at end)
- Added 6 tests for string literal edge cases:
  - Parens inside strings: =IF(A1=")", "yes", "no")
  - Escaped quotes: =CONCATENATE("Say ""hello""", A1)
  - Multiple string literals with parens
  - Unclosed string detection
  - Nested parens with strings

Testing:
- 33 tests in FormattedParsersSpec (32 + 1 precision)
- 23 tests in FormulaInterpolationSpec (17 + 6 string literals)
- All tests passing
- Zero compilation errors

Both fixes maintain Phase 1 minimal validation scope while addressing
real bugs that would reject valid user input.

+++ xl-core/src/com/tjclp/xl/formatted/FormattedParsers.scala
@@ -32,7 +32,7 @@ object FormattedParsers:
       val cleaned = s.replaceAll("[\\$,\\s]", "")
       if cleaned.isEmpty then throw new NumberFormatException("Empty value after removing $ and ,")
       val num = BigDecimal(cleaned)
-      Formatted(CellValue.Number(num.toDouble), NumFmt.Currency)
+      Formatted(CellValue.Number(num), NumFmt.Currency)
     }.toEither.left.map { err =>
       XLError.MoneyFormatError(s, err.getMessage)
     }
@@ -57,7 +57,7 @@ object FormattedParsers:
       val cleaned = s.replace("%", "").trim
       if cleaned.isEmpty then throw new NumberFormatException("Empty value after removing %")
       val num = BigDecimal(cleaned) / 100
-      Formatted(CellValue.Number(num.toDouble), NumFmt.Percent)
+      Formatted(CellValue.Number(num), NumFmt.Percent)
     }.toEither.left.map { err =>
       XLError.PercentFormatError(s, err.getMessage)
     }
@@ -108,7 +108,7 @@ object FormattedParsers:
       val cleaned = s.replaceAll("[\\$,()\\s]", "")
       if cleaned.isEmpty then throw new NumberFormatException("Empty value after removing $ , ( )")
       val num = if isNegative then -BigDecimal(cleaned) else BigDecimal(cleaned)
-      Formatted(CellValue.Number(num.toDouble), NumFmt.Currency)
+      Formatted(CellValue.Number(num), NumFmt.Currency)
     }.toEither.left.map { err =>
       XLError.AccountingFormatError(s, err.getMessage)
     }
